0.1

+++ tasks/mpi/sidorina_p_broadcast/src/ops_mpi.cpp
@@ -30,10 +30,7 @@ bool sidorina_p_broadcast_mpi::RefBroadcast::pre_processing() {
 bool sidorina_p_broadcast_mpi::RefBroadcast::validation() {
   internal_order_test();
   if (world.rank() == 0) {
-    return taskData->inputs_count[0] > 0 && 
-           taskData->inputs_count[1] > 0 && 
-           taskData->outputs_count[0] > 0 && 
-           taskData->inputs_count[0] == taskData->outputs_count[0];
+    return taskData->inputs_count[0] > 0 && taskData->inputs_count[1] > 0 && taskData->outputs_count[0] > 0 && taskData->inputs_count[0] == taskData->outputs_count[0];
   }
   return true;
 }
@@ -41,35 +38,35 @@ bool sidorina_p_broadcast_mpi::RefBroadcast::validation() {
 bool sidorina_p_broadcast_mpi::RefBroadcast::run() {
   internal_order_test();
 
-  int root = 0; 
-  MPI_Bcast(&delta_, 1, MPI_INT, root, world); 
-  MPI_Bcast(&size_, 1, MPI_INT, root, world); 
+  int root = 0;
+  MPI_Bcast(&delta_, 1, MPI_INT, root, world);
+  MPI_Bcast(&size_, 1, MPI_INT, root, world);
 
-  res.resize(size_, 0); 
+  res.resize(size_, 0);
   if (world.rank() != root) {
-    arr.resize(size_); 
+    arr.resize(size_);
   } else {
-    arr.resize(size_); 
+    arr.resize(size_);
     std::copy(term.data(), term.data() + delta_, arr.begin());
   }
 
   MPI_Bcast(arr.data(), size_, MPI_INT, root, world);
 
   if (world.rank() == root) {
-    for (size_t p = 1; p < world.size(); ++p) {
+    for (int p = 1; p < world.size(); ++p) {
       MPI_Send(term.data() + p * delta_, delta_, MPI_INT, p, 0, world);
     }
   } else {
     MPI_Recv(term.data(), delta_, MPI_INT, root, 0, world, MPI_STATUS_IGNORE);
   }
 
-  for (size_t i = 0; i < static_cast<int>(arr.size()); ++i) {
+  for (int i = 0; i < static_cast<int>(arr.size()); ++i) {
     int num = arr[i];
 
     int result = 0;
-    for (size_t t : term) {
+    for (int t : term) {
       if (t >= 0) {
-        result += num+t;
+        result += num + t;
       }
     }
     arr[i] = result;
