+++ tasks/mpi/sidorina_p_broadcast/src/ops_mpi_m.cpp
@@ -10,7 +10,7 @@
 using namespace boost::mpi;
 
 template <typename T>
-inline void sidorina_p_broadcast_mpi::Broadcast::broadcast_m(const boost::mpi::communicator& comm,T& value, int root) {
+inline void sidorina_p_broadcast_mpi::Broadcast::broadcast_m(const boost::mpi::communicator& comm, T& value, int root) {
   int n = comm.size();
   if (n <= 2) {
     if (comm.rank() == root) {
@@ -28,8 +28,8 @@ inline void sidorina_p_broadcast_mpi::Broadcast::broadcast_m(const boost::mpi::c
   std::vector<int> recipients(comm.size());
 
   if (comm.rank() == root) {
-    recipients[ (root + 1) % comm.size()] = value;
-    recipients[ (root + 2) % comm.size()] = value;
+    recipients[(root + 1) % comm.size()] = value;
+    recipients[(root + 2) % comm.size()] = value;
   } else {
     int id_elem = comm.rank() - root;
     if (comm.rank() < root) {
@@ -37,7 +37,6 @@ inline void sidorina_p_broadcast_mpi::Broadcast::broadcast_m(const boost::mpi::c
     }
 
     int id_sender = (root + (id_elem - 1) / 2) % comm.size();
-    
     comm.recv(id_sender, 0, value);
   }
 }
@@ -96,10 +95,7 @@ bool sidorina_p_broadcast_mpi::Broadcast::pre_processing() {
 bool sidorina_p_broadcast_mpi::Broadcast::validation() {
   internal_order_test();
   if (world.rank() == 0) {
-    return taskData->inputs_count[0] > 0 &&
-           taskData->inputs_count[1] > 0 &&
-           taskData->outputs_count[0] > 0 &&
-           taskData->inputs_count[0] == taskData->outputs_count[0];
+    return taskData->inputs_count[0] > 0 && taskData->inputs_count[1] > 0 && taskData->outputs_count[0] > 0 && taskData->inputs_count[0] == taskData->outputs_count[0];
   }
   return true;
 }
@@ -122,20 +118,20 @@ bool sidorina_p_broadcast_mpi::Broadcast::run() {
   MPI_Bcast(arr.data(), size_, MPI_INT, root, world);
 
   if (world.rank() == root) {
-    for (size_t p = 1; p < world.size(); ++p) {
+    for (int p = 1; p < world.size(); ++p) {
       MPI_Send(term.data() + p * delta_, delta_, MPI_INT, p, 0, world);
     }
   } else {
     MPI_Recv(term.data(), delta_, MPI_INT, root, 0, world, MPI_STATUS_IGNORE);
   }
 
-  for (size_t i = 0; i < static_cast<int>(arr.size()); ++i) {
+  for (int i = 0; i < static_cast<int>(arr.size()); ++i) {
     int num = arr[i];
 
     int result = 0;
-    for (size_t t : term) {
+    for (int t : term) {
       if (t >= 0) {
-        result += num+t;
+        result += num + t;
       }
     }
     arr[i] = result;
